fix comments in feat/16446-navigation-validation (#4094)
* fix comments in feat/16446-navigation-validation

* fixed tests in NavigationButtons tests

* refactor: implement navigation with validation logic in Page and PageGroup components

---------
